requirements.txt
aiofiles==23.1.0
aioredis==2.0.1
aiosqlite==0.19.0
alembic==1.11.1
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
asyncpg==0.28.0
bcrypt==4.0.1
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.2
distlib==0.3.7
dnspython==2.4.0
email-validator==2.0.0.post2
et-xmlfile==1.1.0
excel2json-3==0.1.6
exceptiongroup==1.1.2
fastapi==0.100.0
fastapi-users==12.1.0
fastapi-users-db-sqlalchemy==6.0.0
filelock==3.12.2
gevent==23.7.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
identify==2.5.26
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.1
makefun==1.15.1
Mako==1.2.4
MarkupSafe==2.1.3
nodeenv==1.8.0
numpy==1.25.2
openpyxl==3.1.2
orjson==3.9.2
packaging==23.1
pandas==2.0.3
passlib==1.7.4
platformdirs==3.10.0
pluggy==1.2.0
pre-commit==3.3.3
prompt-toolkit==3.0.39
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycparser==2.21
pydantic==2.0.3
pydantic-extra-types==2.0.0
pydantic-settings==2.0.2
pydantic_core==2.3.0
PyJWT==2.7.0
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-dependency==0.5.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
tomli==2.0.1
types-requests==2.31.0.2
types-urllib3==1.26.25.14
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
urllib3==2.0.4
uvicorn==0.23.1
vine==5.0.0
virtualenv==20.24.2
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
xlrd==2.0.1
zope.event==5.0
zope.interface==6.0
